moved getTaskCompletedDays to another file
