+++ scripts/done-task.js
@@ -1,5 +1,6 @@
 import { CLICK_EVENT, HIDE_CLASS } from "/scripts/constants.js";
 import createButton from "/scripts/factory/createButton.js";
+import getTaskCompletedDays from "/scripts/factory/timeToComplete.js";
 
 function onTaskComplete(
   doneButton,
@@ -10,13 +11,6 @@ function onTaskComplete(
   id
 ) {
   doneButton.addEventListener(CLICK_EVENT, doneTask);
-
-  function getTaskCompletedDays(time) {
-    const now = Date.now();
-    const diffTime = now - time;
-    const diffHours = diffTime / (1000 * 60 * 60);
-    return Math.ceil(diffHours / 24);
-  }
 
   function doneTask() {
     inputElement.classList.add("done");
